+++ docs/changelog.md
@@ -18,6 +18,7 @@
 
 Added:
 
+- An [XGM][extra.components.XGM] component to access XGM devices (!53).
 - [PumpProbePulses][extra.components.PumpProbePulses] to combine X-ray and optical laser pulses in a single pattern (!24).
 - The [Scan][extra.components.Scan] component to automatically detect steps
   within a motor scan (!4).